fixes for deck import, issue 1261

+++ src/com/ichi2/libanki/importer/Anki2Importer.java
@@ -422,7 +422,7 @@ public class Anki2Importer {
             		card[8] = (Long) card[8] - aheadBy;
             	}
             	// if odid true, convert card from filtered to normal
-            	if ((Integer)card[15] != 0) {
+            	if ((Long)card[15] != 0) {
             		// odid
             		card[15] = 0;
             		// odue
